+++ CHANGELOG.md
@@ -10,6 +10,18 @@ We appreciate your patience while we speedily work towards a stable release of t
 
 <!-- NEW CONTENT GENERATED BELOW. PLEASE PRESERVE THIS COMMENT. -->
 
+### 0.62.186 (2024-05-29)
+
+* `modal.Volume` background commits is now enabled by default when using `spawn_sandbox`.
+
+
+
+### 0.62.185 (2024-05-28)
+
+- The `modal app stop` CLI command now accepts a `--name` (or `-n`) option to stop an App by name rather than by ID.
+
+
+
 ### 0.62.181 (2024-05-24)
 
 * Background committing on `modal.Volume` mounts is now default behavior.